refactor(pipeline): extract timeout functions to timeout.go [#466]
Move applyPipelineTimeout, wrapTimeoutError, and lastRunningPhase
from engine.go to timeout.go.
